# docs/phase2/29_API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: FlexSpace API
  version: "1.0.0"
  description: >
    FlexSpace Flexible Workspace Marketplace — Phase 2 API contract.
    Implements the decisions in 11_API_CONTRACTS.md. REST/JSON, versioned base path,
    JWT bearer auth, consistent error envelope, idempotency keys on mutating
    financial/booking endpoints. This is a contract, not an implementation —
    no server code exists yet.
servers:
  - url: https://api.flexspace.az/api/v1
    description: Production (placeholder domain)
  - url: https://staging-api.flexspace.az/api/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /auth/otp/request:
    post:
      tags: [Auth]
      summary: Request an OTP code (passwordless auth, 11_API_CONTRACTS.md §11.4)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier: { type: string, description: "email or phone" }
      responses:
        # The shipped AuthController returns 204 (@HttpCode(HttpStatus.NO_
        # CONTENT)), not 200 as this line previously said — a second
        # spec/implementation drift found alongside the missing /auth/logout
        # path above, both while building the frontend's typed client
        # against this file.
        "204": { description: OTP sent }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and receive tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, code]
              properties:
                identifier: { type: string }
                code: { type: string }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        # AuthService also throws these for this route (OTP_TOO_MANY_ATTEMPTS,
        # USER_NOT_FOUND) — added for the same drift-correction reason as
        # the other changes in this section.
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate access token using a refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        # AuthService throws UnauthorizedException (INVALID_REFRESH_TOKEN /
        # ACCOUNT_INACTIVE) for this route — missing from the original spec.
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token (18_SECURITY.md)
      # Documented here to match the shipped AuthController, which has
      # implemented and tested this route (23/23 unit, 79/79 e2e passing)
      # since before this contract file was last touched — a genuine
      # spec/implementation drift, not a new endpoint being proposed.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "204": { description: Refresh token revoked }

  # ------------------------------------------------------------------------
  # PUBLIC / CUSTOMER — SEARCH & ROOMS
  # ------------------------------------------------------------------------
...
  /spaces/{roomId}/availability:
    get:
      tags: [Rooms]
      summary: Live availability for a room/date (12_RESERVATION_ENGINE.md §12.1)
      security: []
      parameters:
        - { $ref: "#/components/parameters/RoomId" }
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
      responses:
        "200":
          description: >-
            Open windows for the given local calendar date, already
            filtered to windows that can fit at least the room's own
            `minBookingMinutes` (BookingsController.getAvailability /
            AvailabilityService.getOpenWindows — takes only `roomId` and
            `date`; a previously-documented required `durationMinutes`
            param was doc-only drift, found and removed while building
            the room detail page's booking widget — the real endpoint
            never read it). Each window may be longer than one booking;
            picking a specific `startAt`/duration within a window, and
            keeping it within `[minBookingMinutes, maxBookingMinutes]`,
            is the caller's job.
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      type: object
                      properties:
                        startAt: { type: string, format: date-time }
                        endAt: { type: string, format: date-time }

  # ------------------------------------------------------------------------
  # BOOKINGS & PAYMENTS
  # ------------------------------------------------------------------------
  /bookings:
    post:
      tags: [Bookings]
      summary: Create a booking hold (DRAFT/PENDING) — 12_RESERVATION_ENGINE.md §12.3
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roomId, startAt, endAt]
              properties:
                roomId: { type: string, format: uuid }
                startAt: { type: string, format: date-time }
                endAt: { type: string, format: date-time }
                participants: { type: integer }
                purpose: { type: string }
                customer:
                  type: object
                  description: >-
                    Not required in the schema — `CreateBookingDto.customer`
                    is `@IsOptional()`. Only conditionally required, and only
                    at runtime: an unauthenticated (guest) caller must supply
                    `email` or `phone` here, or `BookingsService.create`
                    throws `CUSTOMER_IDENTIFIER_REQUIRED` (400). An
                    authenticated caller's JWT already identifies the
                    customer, so this object is ignored either way once a
                    session exists. Fixed as doc-only drift — a previous
                    version of this doc listed `customer` under the
                    top-level `required` array, which doesn't match
                    `CreateBookingDto`.
                  properties:
                    name: { type: string }
                    email: { type: string }
                    phone: { type: string }
      responses:
        "201":
          description: Booking created (PENDING), price breakdown included
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingSummary" }
        "409":
          description: Slot no longer available (race condition — 12_RESERVATION_ENGINE.md §12.2)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /bookings/{bookingId}:
    get:
      tags: [Bookings]
      summary: Get booking detail — guest (by opaque id) or the owning authenticated customer
      security: []
      description: >-
        `@Public()` with an optional bearer token, not a hard auth wall
        (`BookingsController.getOne`). `POST /bookings` is itself public
        (guest checkout, 05_USER_FLOWS.md §5.2) and provisions a user
        account with no JWT issued, so a guest who just booked has no
        session to prove ownership with — the frontend's payment-
        confirmation page (`/booking/{bookingId}/confirming`, driven by
        `payments.successUrlTemplate`) depends on this endpoint working for
        that guest. Possession of the opaque booking UUID is the gate for
        an anonymous caller; a caller who *does* send a valid bearer token
        still gets the ownership check — a non-owning authenticated caller
        gets 404, never revealing that the booking exists. Fixed as doc-only
        drift while building the payment-confirmation page: this endpoint
        previously required auth unconditionally, which would have 401'd
        every guest booking, the primary case `POST /bookings` supports.
      parameters:
        - { $ref: "#/components/parameters/BookingId" }
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingSummary" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /account/bookings:
    get:
      tags: [Account]
      summary: List my bookings
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [upcoming, past, cancelled] } }
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/BookingSummary" } }
...
  /favorites/{roomId}:
    get:
      tags: [Favorites]
      summary: Whether I have favorited this room
      parameters: [{ $ref: "#/components/parameters/RoomId" }]
      responses:
        "200":
          description: Favorite status
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorited: { type: boolean }
    post:
      tags: [Favorites]
      summary: Favorite a room
      description: Idempotent — favoriting an already-favorited room just returns the same result.
      parameters: [{ $ref: "#/components/parameters/RoomId" }]
      responses:
        "201":
          # FavoritesController.add() has no @HttpCode() override and
          # FavoritesService.add() returns a value, so Nest's default for
          # a POST handler applies (201, not the previously-documented
          # bare 204) with the returned JSON body. Fixed as doc-only drift
          # found while wiring the room detail page's bookmark button.
          description: Added (or already favorited)
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorited: { type: boolean, enum: [true] }
    delete:
      tags: [Favorites]
      summary: Unfavorite a room
      description: Idempotent — unfavoriting a room that isn't favorited just returns the same result.
      parameters: [{ $ref: "#/components/parameters/RoomId" }]
      responses:
        "200":
          # Same fix as POST above: no @HttpCode() on remove(), so Nest's
          # default GET/DELETE status (200) applies, not 204, and the
          # response has a JSON body.
          description: Removed (or already not favorited)
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorited: { type: boolean, enum: [false] }
...
  /provider/bookings:
    get:
      tags: [Provider]
      summary: List bookings across my rooms
      parameters:
        - { name: locationId, in: query, schema: { type: string, format: uuid } }
        - { name: roomId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/BookingSummary" } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema: { type: string, format: uuid }
    BookingId:
      name: bookingId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: 11_API_CONTRACTS.md §11.2 — required in practice for booking/payment mutations
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found (or not owned by caller — 11_API_CONTRACTS.md §11.4)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "SLOT_UNAVAILABLE" }
            message: { type: string }
            details: { type: object }
    AuthTokens:
      type: object
      # All three are always present — AuthService.verifyOtp/refreshTokens
      # return the `TokenPair` interface (backend/src/modules/auth/
      # auth.service.ts), which has them as non-optional fields. Marked
      # `required` here to match (previously unmarked, which made the
      # generated frontend client treat them as possibly-undefined).
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: "access token lifetime in seconds" }
...
    BookingSummary:
      type: object
      description: >-
        Matches the real shape returned by `BookingsService.create` /
        `findById` / `listForCustomer` — the raw `BookingEntity` plus its
        `items` relation, not a hand-assembled DTO. Fixed as doc-only
        drift: a previous version of this schema had a single `room` (a
        full `RoomSummary`) and top-level `startAt`/`endAt`, and priced
        `grossAmount`/`serviceFeeAmount`/`totalAmount` as three separate
        `Money` objects. The entity has neither — pricing is one shared
        `currency` column plus three plain integer (minor-unit) amount
        columns, and the room/time range live per line-item on
        `booking_item` (`items[]`), not on the booking itself; a booking
        may in principle cover more than one item/room. `booking_item`
        only stores `roomId`, not a full room summary — the frontend
        already has the room's display details from the room detail page
        it booked from.
      properties:
        id: { type: string, format: uuid }
        customerUserId: { type: string, format: uuid }
        status: { type: string, enum: [DRAFT, PENDING, PAYMENT_PENDING, CONFIRMED, COMPLETED, CANCELLED, EXPIRED, NO_SHOW, REFUND_PENDING, REFUNDED] }
        currency: { type: string, example: "AZN" }
        grossAmount: { type: integer, description: "minor units (qəpik)" }
        serviceFeeAmount: { type: integer, description: "minor units (qəpik)" }
        totalAmount: { type: integer, description: "minor units (qəpik)" }
        purpose: { type: string, nullable: true }
        participantsCount: { type: integer, nullable: true }
        holdExpiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              roomId: { type: string, format: uuid }
              startAt: { type: string, format: date-time }
              endAt: { type: string, format: date-time }
              unitPriceAmount: { type: integer, description: "minor units (qəpik)" }
              quantity: { type: integer }
              status: { type: string, enum: [DRAFT, PENDING, PAYMENT_PENDING, CONFIRMED, COMPLETED, CANCELLED, EXPIRED, NO_SHOW, REFUND_PENDING, REFUNDED] }
...